fix: 핸드폰 번호 포맷 로직 이동
