+++ src/features/user/components/login/index.tsx
@@ -1,4 +1,4 @@
-import { useCallback, useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
 
 import { useToast } from "~/app/providers/ToastProvider";
@@ -14,7 +14,7 @@ const Login = () => {
 	const [isSaved, setIsSaved] = useState(false);
 
 	const navigate = useNavigate();
-	const { showToast } = useToast();
+	const showToast = useToast().showToast;
 
 	useEffect(() => {
 		const savedPhoneNumber = localStorage.getItem(PHONE_NUMBER_STORAGE_KEY);
@@ -26,29 +26,12 @@ const Login = () => {
 
 	const isPhoneNumberValid = phoneNumber.length === 13;
 
-	const formatPhoneNumber = useCallback((value: string) => {
-		const numbers = value.replace(/[^0-9]/g, "");
-
-		if (numbers.length <= 3) return numbers;
-		if (numbers.length <= 7) return `${numbers.slice(0, 3)}-${numbers.slice(3)}`;
-		return `${numbers.slice(0, 3)}-${numbers.slice(3, 7)}-${numbers.slice(7, 11)}`;
-	}, []);
-
-	const handlePhoneNumberChange = useCallback(
-		(e: React.ChangeEvent<HTMLInputElement>) => {
-			const formattedValue = formatPhoneNumber(e.target.value);
-			setPhoneNumber(formattedValue);
-		},
-		[formatPhoneNumber],
-	);
-
 	const handleLogin = () => {
-		showToast(`${phoneNumber}로 로그인 했어요.`, 2000);
-
 		if (isSaved) {
 			localStorage.setItem(PHONE_NUMBER_STORAGE_KEY, phoneNumber);
 		}
 
+		showToast(`${phoneNumber}로 로그인 했어요.`, 2000);
 		navigate("/calculator");
 	};
 
@@ -66,9 +49,9 @@ const Login = () => {
 
 	return (
 		<S.Container>
-			<PhoneNumberInput onChange={handlePhoneNumberChange} value={phoneNumber} />
+			<PhoneNumberInput onChange={setPhoneNumber} value={phoneNumber} />
 			<Button onClick={handleLogin} disabled={!isPhoneNumberValid}>
-				로그인
+				자재 소요 계산하기
 			</Button>
 			<S.Checkbox onClick={handleToggleCheckbox}>
 				<input type="checkbox" checked={isSaved} disabled={!isPhoneNumberValid} />
