+++ src/features/user/components/login/phoneNumberInput/index.tsx
@@ -1,4 +1,5 @@
 import { css } from "@emotion/react";
+import { useCallback } from "react";
 
 import { Input } from "~/components/ui";
 
@@ -9,16 +10,29 @@ const phoneNumberInputStyle = css`
 `;
 
 type Props = {
-	onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+	onChange: (value: string) => void;
 	value: string;
 };
 
 const PhoneNumberInput = ({ onChange, value }: Props) => {
+	const formatPhoneNumber = useCallback((value: string) => {
+		const numbers = value.replace(/[^0-9]/g, "");
+
+		if (numbers.length <= 3) return numbers;
+		if (numbers.length <= 7) return `${numbers.slice(0, 3)}-${numbers.slice(3)}`;
+		return `${numbers.slice(0, 3)}-${numbers.slice(3, 7)}-${numbers.slice(7, 11)}`;
+	}, []);
+
+	const handlePhoneNumberChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+		const formattedValue = formatPhoneNumber(e.target.value);
+		onChange(formattedValue);
+	};
+
 	return (
 		<Input
 			css={phoneNumberInputStyle}
 			value={value}
-			onChange={onChange}
+			onChange={handlePhoneNumberChange}
 			placeholder="핸드폰 번호를 입력해 주세요."
 			inputMode="numeric"
 			maxLength={13}
